+++ src/widgets/Ecommerce.php
@@ -51,7 +51,7 @@ class Ecommerce extends \craft\base\Widget
     /**
      * @inheritDoc IWidget::getBodyHtml()
      *
-     * @return string|false
+     * @return string|null
      * @throws \Twig\Error\LoaderError
      * @throws \yii\base\Exception
      * @throws \yii\base\InvalidConfigException