+++ src/widgets/Realtime.php
@@ -63,7 +63,7 @@ class Realtime extends \craft\base\Widget
     /**
      * @inheritDoc IWidget::getBodyHtml()
      *
-     * @return string|false
+     * @return string|null
      * @throws \Twig\Error\LoaderError
      * @throws \yii\base\Exception
      * @throws \yii\base\InvalidConfigException